Move the config across too, and tidy up the namespacing
