+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sns/SNSClientModule.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.messaging.finatra.modules
+package uk.ac.wellcome.messaging.sns
 
 import javax.inject.Singleton
