+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sns/SNSConfig.scala
@@ -1,3 +1,3 @@
-package uk.ac.wellcome.models.aws
+package uk.ac.wellcome.messaging.sns
 
 case class SNSConfig(topicArn: String)
