+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sqs/SQSConfig.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.models.aws
+package uk.ac.wellcome.messaging.sqs
 
 import scala.concurrent.duration.Duration
