+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/sqs/SQSReaderModule.scala
@@ -1,10 +1,8 @@
-package uk.ac.wellcome.messaging.finatra.modules
+package uk.ac.wellcome.messaging.sqs
 
 import com.amazonaws.services.sqs.AmazonSQS
 import com.google.inject.{Provides, Singleton}
 import com.twitter.inject.TwitterModule
-import uk.ac.wellcome.models.aws.SQSConfig
-import uk.ac.wellcome.messaging.sqs.SQSReader
 
 object SQSReaderModule extends TwitterModule {
